+++ bot/ui/src/utils.js
@@ -1,6 +1,23 @@
 import React from "react";
 
 /**
+ * Generate a browser-originated download.
+ * @param {*} contentType
+ * @param {*} content
+ */
+export function download(contentType, content, filename) {
+  var element = document.createElement('a');
+  element.setAttribute('href', `data:${contentType};charset=utf-8,` + encodeURIComponent(content));
+  element.setAttribute('download', filename);
+
+  element.style.display = 'none';
+  document.body.appendChild(element);
+
+  element.click();
+  document.body.removeChild(element);
+}
+
+/**
  * Format duration in a human-readable way.
  * @param {*} duration
  */